feat: Better error handling and display.
Given cobra also tries to handle and display errors, along with the
usage of the command when we hit an error, this requires special care
and handling.

We disable displaying usage and handling errors in cobra once we are
ready to begin execution of the actual command after all the parsing and
validtion of commands/flags is complete.

The main program still receives both cobra and non-cobra errors emitted
during the homelab execution. The ones from homelab are now wrapped
using a custom error type which allows the main program to only display
those errors and ignore the rest. Hence it is important that the main
program only invokes the cobra homelab command and nothing else unless
we wrap those errors also using the custom homelab runtime error type.